feature(homepage): create homepage
