+++ src/App.css
@@ -93,6 +93,36 @@ button:hover {
   padding: 0 18px;
 }
 
+.username-form-container{
+  width: 100vw;
+  height: 100vh;
+  display: grid;
+  justify-content: center;
+  align-content: center;
+}
+
+.grid-center{
+  justify-self: center;
+  align-self: center;
+}
+
+.username-form {
+  display: grid;
+  width: 500px;
+  max-width: 600px;
+  height: 500px;
+  background: lightblue;
+  border-radius: 5px;
+  flex-direction: column;
+  justify-content: center;
+  align-content: center;
+}
+
+.chat-btn{
+  width: 70%;
+  background-color: yellowgreen;
+}
+
 @media screen and (max-width: 480px) {
   #wrap {
     width: 100%;
