+++ src/components/NotFoundPage/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const NotFound = () => (
+    <div style={{textAlign: "center"}}>
+        <h1>Page not found</h1>
+    </div>
+)
+
+export default NotFound;
